+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 - `[base]` Added `shappoli:active_self` power type. This power works similar to the `apoli:active_self`, but allows specifying multiple keys or key categories. If none of them were specified, action will be triggered if any key was pressed
 - `[trinkets]` Added `shappoli:powers` component for trinkets. This component works similar to the `apoli:powers` component, but works with trinkets
     - TODO: rename to `shappoli:trinket_powers`
+- `[trinkets]` Added `shappoli:has_trinket_power` item condition. This condition checks if the item has a trinket power
 - `[trinkets]` Added `shappoli:equipped_trinket` item condition. This condition checks if the item is equipped by the player as a trinket
 - `[trinkets]` Added `shappoli:slot_linked_active` power type and `slot_linked_key` data type. This power type allows creating active powers assigned to specific trinkets; a key for the power to activate is specified by the slot where the trinket is equipped
 - `[walkers]` Added `shappoli:execute_shape_ability` entity action type. This action type allows executing a specific shape ability